+++ python/python/glide/async_commands/sorted_set.py
@@ -220,12 +220,17 @@ def separate_keys(
     return key_list, weight_list
 
 
-def _create_z_cmd_store_args(
-    destination: str,
+def _create_zinter_zunion_cmd_args(
     keys: Union[List[str], List[Tuple[str, float]]],
     aggregation_type: Optional[AggregationType] = None,
+    destination: Optional[str] = None,
 ) -> List[str]:
-    args = [destination, str(len(keys))]
+    args = []
+
+    if destination:
+        args.append(destination)
+
+    args.append(str(len(keys)))
 
     only_keys, weights = separate_keys(keys)